+++ src/scenes/search/components/QueryBuilder/QueryBuilder.js
