+++ src/scenes/search/components/QueryBuilder/RuleGroup.js
