+++ src/scenes/search/simple/index.css
@@ -27,7 +27,7 @@
 }
 
 .search .home .activities h4 {
-  background: url("./../../assets/002-time.png") left center no-repeat;
+  background: url("../../../assets/002-time.png") left center no-repeat;
   -webkit-background-size: 18px;
   padding-left: 25px;
   background-size: 18px;
